requirements.txt
python-dotenv==1.0.1
requests
apify-client
pymongo[srv]==4.8.0

pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1

langdetect==1.0.9
regex==2024.7.24

transformers==4.44.2
datasets==2.21.0
accelerate==0.33.0
torch>=2.2.0
matplotlib==3.9.2

google-genai